--- likelion13/categories/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class CategoriesConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'categories'
