--- likelion13/categories/models.py
@@ -1,9 +1,0 @@
-from django.db import models
-
-# Create your models here.
-class Category(models.Model):
-    id = models.AutoField(primary_key=True)
-    name = models.CharField(max_length=30)
-
-    def __str__(self):
-        return self.name
