+++ likelion13/posts/models.py
@@ -1,7 +1,5 @@
 from django.db import models
 from accounts.models import User
-from categories.models import Category
-
 # Create your models here.
 # 추상 클래스 정의
 class BaseModel(models.Model):
@@ -11,7 +9,13 @@ class BaseModel(models.Model):
     class Meta:
         abstract = True
 
+class Category(models.Model):
+    id = models.AutoField(primary_key=True)
+    name = models.CharField(max_length=30)
 
+    def __str__(self):
+        return self.name
+    
 class Post(BaseModel):
 
     CHOICES = (
@@ -24,10 +28,16 @@ class Post(BaseModel):
     content = models.TextField()
     status = models.CharField(max_length=15, choices=CHOICES, default='STORED')
     user = models.ForeignKey(User, on_delete=models.CASCADE, related_name='post')
-    category = models.ManyToManyField(Category, blank=True, related_name='post') # 다대다 관계이므로 post가 삭제되어도 category는 유지
 
     def __str__(self): # 표준 파이썬 클래스 메서드, 사람이 읽을 수 있는 문자열을 반환하도록 함
         return self.title
+    
+class PostCategory(models.Model):
+    post = models.ForeignKey(Post, on_delete=models.CASCADE)
+    category = models.ForeignKey(Category, on_delete=models.CASCADE)
+
+    def __str__(self):
+        return f'{self.post.title} - {self.category.name}'
     
 class Comment(BaseModel):
     id = models.AutoField(primary_key=True)
